Database addition
the write to database is working but when it comes to reading from the
database I got it to output into an adapter but not the format i wanted
so I created a custom adapter to display it. But it wasn't working. It
says something about it being null. WHat?!?...This is Day 2.(Defeated
reprise)