BarCode Scanner db addition feature
once a barcode is scanned it is added to the database.and can be seen in
the register tab with the time it was scanned and also the date.but does
not cater to duplicate entries. Nonetheless it was success. p.s
Commiting to master a champion, Feels good when your code works.