Features of the Featureless
Added the ability to export the register to text file it show a number
of attributes in it such as total number of students,course name and
code.also some layout were fixed and adjusted to work smoothly.the
action bar has the words on it(lol if you only knew the reason it wasn't
showing).the scanner also does not save the same ID twice in the
register.Two different types of vibration is used a short one for if the
ID was already scanned and a long one for new ID's.Fixed minor bugs when
the back button and the upbutton was used at different times.Added
dialog for when you are about to export the tables data.Added an about
section for desciption on the developers. I wonder if this is the end of
development.